Fix spliting of LTWA lines and detect ‘eng’ and ‘mul’
The LTWA file is tab separated but we were splitting on any whitespace causing words with spaces in to be split incorrectly and making it harder to detect the languages.

Now we correctly split the fields by tab and then split the language field by comma and strip the whitespace.

We also now accept ‘mul’ as a language which stands for multiple and contains many english abbreviations.