Use a prefilter of the word stem for abbreviator
There was a significant performance penalty in trying every abbreviation regex for every word.

Now we build a dict of dicts where the first level prefilters by the stem of the word (first 3 letters, or less if the word is shorter).

We also return immediatelty for 1 letter words because these cannot be abbreviated.