Switch to urllib and crossref for lookups
We don’t need advanced http features for the lookups. Infact, passing headers such as a user-agent as we are supposed to slows down the lookup. Therefore it’s best to use the standard python urllib rather than the external requests module. This means blib has no requirements outside of the standard library.

I’ve also switched to crossref instead of doi.org. The json format is a tiny bit different but I think crossref resolves faster. The RESTful API is also better documented and has features for searching.