Change cite keys to more unique and easier search order
We usually want to autocomplete in our editor using the authors name rather than the year, so this should come first. I've also switched to using the first page number (instead of issue) as this is more unique incase an author with the same family name publishes in the same issue.