Give alternatives to LTWA abbreviations with diacritics
The LTWA contains some oddities where a word is defined in multiple languages including English but some diacritic is included which is not part of the English, for example:

  nat̡ional-  natl.   rum, fre, eng

We work around this by putting a second entry into the dictionary which is the unicode conversion to standard ascii which will remove the diacritics. It's not entirely clear how safe this is. I tried to use regex to make diacritic letters a choice with their ascii version but it seems impossible to loop over some letter with combining characters (i.e. to treat t̡ as  a single character) regardless of the unicode normalisation.