Handle DOI type errors
Sometimes our DOI list will include valid DOIs which are not a type we can handle (currently anything that is not a journal article).

Previously we raised a RuntimeError, but that completely kills the program with no output, even if we are processing a long list of DOIs.

Now we raise a custom type DoiTypeError and simply print an error message and carry on.