Improve encode utf-8 punctuation support
Use the third party 'regex' module rather than pythons native 're' module because it has more advanced support for unicode. Specifically it supports the \p{} syntax (https://www.regular-expressions.info/unicode.html) which allows us to identify unicode letter and punctuation more reliably. This is important for example when an author name has diacritics. In unicode these may use 'combining marks' which depending on the formatting can come before or after the letter, for example U+0306 is a breve '‘̆' which will appear after a letter such as in 's̆'. Without using \p{} syntax it's extemely difficult to capture the breve with the letter. With the \p{} syntax we can use [\p{L}\p{M}] which captures letters and combining marks on letters.