Add fallback for pylatexenc unknown characters
For unknown characters we now ask pylatexenc to throw an exception. We can then try to handle characters we know how to deal with manually. This still needs some work, because we ideally want to split the string and deal with the character and then process the rest with pylatexenc.